requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
supabase
mangum
requests
